+++ src/leoUI.ts
@@ -5018,6 +5018,13 @@ export class LeoUI extends NullGui {
 
     }
 
+    private _get_focus(): string {
+        const c = g.app.windowList[this.frameIndex].c;
+        const w = g.app.gui.get_focus(c);
+        const focus = g.app.gui.widget_name(w);
+        return focus;
+    }
+
     /**
      * Reveals and selects the specific nav entry in the results of the nav pane.
      */
@@ -5036,11 +5043,14 @@ export class LeoUI extends NullGui {
         void this.leoGotoProvider.navigateNavEntry(p_nav);
     }
 
-    private _get_focus(): string {
+    /**
+     * Call the Leo do_arrow function with that char parameter, then refresh the find panel settings.
+     */
+    public doArrow(key: "Up" | "Down") {
         const c = g.app.windowList[this.frameIndex].c;
-        const w = g.app.gui.get_focus(c);
-        const focus = g.app.gui.widget_name(w);
-        return focus;
+        const fc = c.findCommands;
+        fc.do_arrow(key);
+        this.loadSearchSettings();
     }
 
     /**
@@ -5523,6 +5533,7 @@ export class LeoUI extends NullGui {
         const leoISettings = c.findCommands.ftm.get_settings();
         const w_settings: LeoSearchSettings = {
             // Nav options
+            frozen: scon.frozen,
             navText: scon.navText,
             showParents: scon.showParents,
             isTag: scon.isTag,
@@ -5579,6 +5590,7 @@ export class LeoUI extends NullGui {
         // convert to LeoGuiFindTabManagerSettings
         const searchSettings: LeoGuiFindTabManagerSettings = {
             // Nav settings
+            frozen: p_settings.frozen,
             is_tag: p_settings.isTag,
             nav_text: p_settings.navText,
             show_parents: p_settings.showParents,
@@ -5607,6 +5619,7 @@ export class LeoUI extends NullGui {
 
         // * Try to set the search settings
         // nav settings
+        scon.frozen = searchSettings.frozen;
         scon.navText = searchSettings.nav_text;
         scon.showParents = searchSettings.show_parents;
         scon.isTag = searchSettings.is_tag;
@@ -5673,12 +5686,10 @@ export class LeoUI extends NullGui {
         const suboutlineOnly = searchSettings.suboutline_only || false;
 
         if (!nodeOnly && !suboutlineOnly && !fileOnly) {
-            find.entire_outline = true;
             if (!w.isChecked()) {
                 w.toggle();
             }
         } else {
-            find.entire_outline = false;
             if (w.isChecked()) {
                 w.toggle();
             }
@@ -5928,14 +5939,16 @@ export class LeoUI extends NullGui {
             }
 
             // Either way, try to open the file
-            await this.openLeoFile(vscode.Uri.file(w_result));
+            return this.openLeoFile(vscode.Uri.file(w_result)).then(
+                () => {
+                    // Now, maybe there is a file opened?
+                    if (g.app.windowList.length && g.app.windowList[this.frameIndex]) {
+                        // Already opened file
+                        const c = g.app.windowList[this.frameIndex].c;
+                        g.doHook("recentfiles2", { c: c, p: c.p, v: c.p.v, fileName: w_result });
+                    }
+                });
 
-            // Now, maybe there is a file opened?
-            if (g.app.windowList.length && g.app.windowList[this.frameIndex]) {
-                // Already opened file
-                const c = g.app.windowList[this.frameIndex].c;
-                g.doHook("recentfiles2", { c: c, p: c.p, v: c.p.v, fileName: w_result });
-            }
         }
         return Promise.resolve(undefined);